refactor(interface-transport): split driver-identity contracts into driver-contracts.ts
TDriverId + OWNER/AGENT driver constants + ISubmitOptions move out of
session-contracts.ts together with the CMD-004 driver-routed events
(IUiIntentEvent, ISessionRenamedEvent) — cohesion + the file-size ratchet
(session-contracts may not grow). Legacy-union comments reworded to LEGACY
(the no-deprecated-markers scan bans the @deprecated tag). SPEC synced.
