+++ packages/agent-interface-transport/src/session-contracts.ts
@@ -24,9 +24,14 @@ import type {
   ICommandListEntry,
   ICommandResult,
   TCommandInvocationSource,
-  TCommandUiIntent,
 } from './command-contracts.js';
 import type { ICompactEvent } from './compact-contracts';
+import type {
+  ISessionRenamedEvent,
+  ISubmitOptions,
+  IUiIntentEvent,
+  TDriverId,
+} from './driver-contracts.js';
 import type {
   IContextReferenceItem,
   IMemoryEvent,
@@ -233,24 +238,6 @@ export type TInteractivePermissionHandler = (
  * first already answered.
  */
 
-/**
- * REMOTE-014 E5 co-drive attribution: a stable, SERVER-ASSIGNED id for the driver of an input/turn. It is
- * DISPLAY/ATTRIBUTION ONLY — never an authorization input (the OWNER PRINCIPLE, REMOTE-006, governs
- * authorization; remote == local). Remote = the E3 `deviceId`; local = {@link OWNER_DRIVER_ID}; an
- * agent-wakeup/goal turn = {@link AGENT_DRIVER_ID}.
- */
-export type TDriverId = string;
-
-/** The local operator ("owner") driver id — the default for a human turn with no explicit driver. */
-export const OWNER_DRIVER_ID: TDriverId = 'owner';
-/** The reserved driver id for an autonomous (wakeup/goal/agent-initiated) turn — never the owner. */
-export const AGENT_DRIVER_ID: TDriverId = 'agent';
-
-/** REMOTE-014 E5: options for `submit` — carries the SERVER-ASSIGNED driver id for co-drive attribution. */
-export interface ISubmitOptions {
-  readonly driverId?: TDriverId;
-}
-
 /** A tool call awaiting a permission decision. Serializable — crosses the transport boundary unchanged. */
 export interface IPermissionRequestEvent {
   id: string;
@@ -273,27 +260,6 @@ export interface IPromptResolvedEvent {
   id: string;
   /** REMOTE-014 E5: the driver who answered the prompt (server-assigned; display-only). */
   answererDriverId?: TDriverId;
-}
-
-/**
- * CMD-004 Phase 2: a command-issued UI intent, emitted as a fire-and-forget `ui_intent` session
- * event. Routed to the REQUESTING surface: `requesterDriverId` is stamped from the command-origin
- * driver id passed into `executeCommand` (the REMOTE-014 E5 server-assigned id for remote surfaces;
- * the active turn's driver only as a fallback for model-invoked commands). Other surfaces ignore it;
- * an intent needs no answer (no parking, no response promise). Serializable.
- */
-export interface IUiIntentEvent {
-  intent: TCommandUiIntent;
-  /** The server-assigned driver id of the surface that issued the command (routing/display-only). */
-  requesterDriverId?: TDriverId;
-}
-
-/**
- * CMD-004 Phase 2: the session was renamed (host-executed `session-rename` action). Broadcast so
- * every attached surface — including co-driving ones — updates its title. Serializable.
- */
-export interface ISessionRenamedEvent {
-  name: string;
 }
 
 /** Emitted when a context file is found stale and re-read before a turn. */
